refactor(AppState): split five more logic clusters out of AppState.swift
AppState.swift: 819 -> 581 lines (-30%). New files: AppState+Lifecycle.swift
(63, one-shot init/scene/onboarding entry points), AppState+PushRouting.swift
(62, iOS-only custom-push tap-routing types + popup shown/seen bookkeeping,
macOS-excluded like AppState+LiveActivity.swift), AppState+SyncPresentation.swift
(40, SyncConflictItem/SyncSource + isSyncLocalOnly), AppState+SettingsDerived.swift
(47, accentColor/savedAnnouncementDepartments/visualStylePolicy), and
AppState+CourseRelabeling.swift (116, the relabelAllCachedCourses sweep -
the single biggest extraction).

Two access-control widenings, both from the relabel extraction:
- `relabelAllCachedCourses`: private -> internal. Three didSet observers
  that must stay on the class (useEnglishCourseAbbreviation,
  useEnglishClassroomAbbreviation, classroomMandarinDisplay) call it, and
  Swift's `private` is file-scoped.
- `relabelTask`: private -> internal. It's a stored property so it can't
  move; its only reader/writer now lives in a different file.
No other `private` member in AppState.swift needed widening: the four
NotificationCenter observer vars and widgetReloadCoordinator are only
touched from init()/deinit() (unmoved), SyncConflictItem's private
statusLabel moved with its enclosing struct as one unit, and
isSyncLocalOnly only *reads* lastSyncSource (private(set) restricts the
setter, not the getter) so it needed no widening to move.

Splitting the custom-push-routing #if os(iOS) block into a stored-property
half (stays) and a types/methods half (moves) means each half now needs
its own #if os(iOS)/#endif to preserve the original compile-time gating -
net +1 `#if os(iOS)` and +1 `#endif` versus the original single pair.
That is the only line-diff beyond the previously-established pure-move
exceptions (private->internal, import lines, `extension AppState {`,
closing braces, file-header doc comments).

Did not reach "well under 500": init() (~149 lines, the sole/designated
initializer - Swift only allows convenience inits in extensions) and
deinit() (~21 lines, categorically forbidden in extensions) can't move
without either breaking the language rules or rewriting init's body into
helper-method calls, which would cross the pure-move line. The rest of
the file is stored-property declarations (many with inline didSet/
initializer closures that can't be split from the declaration) that the
Observation macro and "extensions can't hold stored properties" both
require to stay in the class body. Left `static let themeColors` (10
lines) alone too - Swift extensions technically can hold static stored
properties, but the task's stated rule doesn't carve out that exception
and it wouldn't have closed the gap regardless. 581 lines appears to be
at or near the real floor under these constraints.

Verified: tools/check_macos_sources.py passes (2 allowlists in
agreement). macOS Debug build succeeded. iOS build succeeded via
`-destination 'generic/platform=iOS Simulator'` (Debug) - no simulator
runtimes are installed in this environment, so the requested booted
iPhone 17 Pro destination and every other iPhone simulator are
"unavailable, runtime profile not found"; this is the closest available
substitute. Pure-move check: sorted non-blank-line multiset of
`git show HEAD~1:swift/TigerDuck/App/AppState.swift` vs. the new
AppState.swift + all five new files matches except for the two
widenings, the duplicated `#if os(iOS)`/`#endif` pair, and the expected
import/extension/brace/header-comment lines.
